Refactor error page: move content from glitch page to main page and remove glitch page
